softdesk/views/project.py
from softdesk.models import Project
from softdesk.serializers import ProjectSerializer
from rest_framework import generics


class ProjectListView(generics.ListCreateAPIView):
    queryset = Project.objects.all()
    serializer_class = ProjectSerializer


class ProjectDetailView(generics.RetrieveUpdateDestroyAPIView):
    queryset = Project.objects.all()
    serializer_class = ProjectSerializer



